[CC] comp_loudness_alt => comp_loudness
