+++ validations/loudness_ecma/hearing_model_validation.py
@@ -5,7 +5,7 @@ import matplotlib.pyplot as plt
 from scipy.optimize import root_scalar
 
 # Project Imports
-from mosqito.functions.loudness_ecma.comp_loudness_alt import comp_loudness
+from mosqito.functions.loudness_ecma.comp_loudness import comp_loudness
 from mosqito.functions.loudness_ecma.equal_loudness_contours import (
     equal_loudness_contours,
 )
